Refactor calculator component with hooks
