+++ src/components/Calculator.js
@@ -1,54 +1,46 @@
-import React from 'react';
+import React, { useState } from 'react';
 import './caclulator.css';
 import calculate from '../logic/calculate';
 import Button from './Button';
 
-class Calculator extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this.state = {
-      total: null,
-      next: null,
-    };
-    this.handleButtonClick = this.handleButtonClick.bind(this);
-  }
+function Calculator() {
+  const [state, setState] = useState({ total: null, next: null, operation: null });
 
-  handleButtonClick(char) {
-    this.setState((state) => (calculate(state, char)));
-  }
+  const handleButtonClick = (char) => {
+    setState(calculate(state, char));
+  };
 
-  setDisplay(content) {
-    this.setState(() => ({ next: content }));
-  }
+  const setDisplay = (content) => {
+    state.next = content;
+  };
 
-  render() {
-    const { next, total } = this.state;
-
-    return (
-      <div className="calculator">
-        <input type="text" value={next || total || ''} onChange={(e) => this.setDisplay(e.target.value)} />
-        <Button buttonValue="AC" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="%" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="+/-" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="÷" buttonClass="orange" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="7" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="8" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="9" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="x" buttonClass="orange" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="4" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="5" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="6" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="-" buttonClass="orange" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="1" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="2" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="3" buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="+" buttonClass="orange" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="0" buttonClass="silver btn0" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="." buttonClass="silver" handleButtonClick={this.handleButtonClick} />
-        <Button buttonValue="=" buttonClass="orange" handleButtonClick={this.handleButtonClick} />
-      </div>
-    );
-  }
+  const buttons = [
+    { value: 'AC', class: 'silver' },
+    { value: '+/-', class: 'silver' },
+    { value: '%', class: 'silver' },
+    { value: '÷', class: 'orange' },
+    { value: '7', class: 'silver' },
+    { value: '8', class: 'silver' },
+    { value: '9', class: 'silver' },
+    { value: 'x', class: 'orange' },
+    { value: '4', class: 'silver' },
+    { value: '5', class: 'silver' },
+    { value: '6', class: 'silver' },
+    { value: '-', class: 'orange' },
+    { value: '1', class: 'silver' },
+    { value: '2', class: 'silver' },
+    { value: '3', class: 'silver' },
+    { value: '+', class: 'orange' },
+    { value: '0', class: 'silver btn0' },
+    { value: '.', class: 'silver' },
+    { value: '=', class: 'orange' },
+  ];
+  return (
+    <div className="calculator">
+      <input type="text" value={state.next || state.total || ''} onChange={(e) => setDisplay(e.target.value)} />
+      {buttons.map((btn) => <Button key="1" buttonValue={btn.value} buttonClass={btn.class} handleButtonClick={handleButtonClick} />)}
+    </div>
+  );
 }
 
 export default Calculator;
